refactor: search into smaller files and add types
